Add parameter to enable overriding tree args
Adds a configuration parameter to the workflow, `override_default_args`,
in the `tree` section that allows users to toggle the corresponding new
flag in `augur tree` on or off. Sets the default value to `true`, since
the custom tree builder arguments we define in defaults are redundant
with previously hardcoded defaults in augur tree. This change should fix
a related segmentation fault on some systems, but it also allows users
to define more flexible tree builder arguments that conflict with the
defaults (e.g., bootstrap mode in IQ-TREE).